refactor: centralize anti raid core logic and execution
