+++ packages/yuudachi/locales/en-US/translation.json
@@ -130,7 +130,8 @@
 							"reject_self": "This user is you.",
 							"reject_protected": "This user has a role ignored by the Anti-Raid-Nuke.",
 							"reject_user": "This user is not in the guild."
-						}
+						},
+						"no_hits": "The Anti-Raid-Nuke will not hit anyone, you may want to re-adjust the parameters and try again."
 					},
 					"parameters": {
 						"heading": "Parameters:",
@@ -138,7 +139,8 @@
 						"dry_run": "• Operating in **dry-run** mode, no members will be banned",
 						"days_zero": "• Not deleting any messages",
 						"days_one": "• Deleting messages sent in the last day",
-						"days_other": "• Deleting messages sent in the last `{{count}}` days"
+						"days_other": "• Deleting messages sent in the last `{{count}}` days",
+						"parsed_ids": "• Valid members of this guild: `{{valid}}` of `{{total}}` total parsed ids."
 					},
 					"reason": "Anti-Raid-Nuke `({{current}}/{{count}})`",
 					"pending": "Anti-Raid-Nuke is projected to hit `{{count}}` members.\n• Created within `{{creation_range}}`\n• Joined within `{{join_range}}`",
@@ -146,14 +148,12 @@
 						"execute": "Launch",
 						"dry_run": "Dry run"
 					},
-					"attachment": "{{user_id}} | Join: {{- joined_at}} | Creation: {{- created_at}} | {{- user_tag}}`",
 					"cancel": "Canceled Anti-Raid-Nuke.",
 					"success": "Anti-Raid-Nuke hit `{{count}}` member."
 				},
 				"filter": {
 					"errors": {
-						"no_hits": "The Anti-Raid-Nuke will not hit anyone, you might want to re-adjust the parameters and try again.",
-						"invalid_avatar": "Invalid avatar format, must be an raw hash, a user id, a URL or `none`",
+						"invalid_avatar": "Invalid avatar format, must be a raw hash, user id, URL or `none`",
 						"pattern_format": "Invalid pattern provided."
 					},
 					"parameters": {
@@ -167,32 +167,28 @@
 						},
 						"pattern": "• Username matching pattern:",
 						"avatar": "• Avatar hash (user or member) matching: `{{- avatar}}`",
-						"zalgo": "• Filtering usernames with Zalgo",
-						"confusables": "• Filtering usernames with Confusables"
+						"zalgo": "• Usernames including Zalgo characters",
+						"filter_confusables": "• Usernames including confusables",
+						"clean_confusables": "• Cleaning usernames including confusables before applying other filters"
 					}
 				},
 				"file": {
 					"errors": {
-						"no_hits": "The Anti-Raid-Nuke will not hit anyone, the file you provided did not match any users.",
 						"no_ids": "This file does not contain any ids."
 					},
 					"parameters": {
-						"file": "• File: {{- file}}",
-						"users": "• Out of all of the ids, `{{count}}` aren't members of this guild"
+						"file": "• File: {{- file_link}}",
+						"link_label": "<CDN url>"
 					}
 				},
 				"modal": {
 					"errors": {
-						"no_hits": "The Anti-Raid-Nuke will not hit anyone, the content you provided did not match any users.",
 						"no_ids": "The content does not contain any ids."
-					},
-					"parameters": {
-						"users": "• Out of all of the ids, `{{count}}` aren't members of this guild"
 					},
 					"title": "Anti-Raid-Nuke Entry",
 					"components": {
 						"label": "{{i}}- Enter valid ids",
-						"placeholder": "Enter ids separated by spaces or any other valid separator (e.g. `,|/`)"
+						"placeholder": "Members will be parsed by id from the provided content."
 					}
 				}
 			},
