+++ packages/yuudachi/src/commands/moderation/sub/anti-raid-nuke/filter.ts
@@ -1,11 +1,8 @@
-import { Buffer } from 'node:buffer';
 import dayjs from 'dayjs';
-import { ButtonStyle, codeBlock, ComponentType, time, TimestampStyles } from 'discord.js';
+import { codeBlock, time, TimestampStyles } from 'discord.js';
 import i18next from 'i18next';
-import { nanoid } from 'nanoid';
+import { handleAntiRaidNuke } from './coreCommand.js';
 import type { InteractionParam, ArgsParam, LocaleParam } from '../../../../Command.js';
-import { DATE_FORMAT_LOGFILE } from '../../../../Constants.js';
-import { blastOff } from '../../../../functions/anti-raid/blastOff.js';
 import {
 	ageFilter,
 	joinFilter,
@@ -14,15 +11,9 @@ import {
 	patternFilter,
 	zalgoFilter,
 } from '../../../../functions/anti-raid/filters.js';
-import { formatMemberTimestamps } from '../../../../functions/anti-raid/formatMemberTimestamps.js';
 import { parseAvatar } from '../../../../functions/anti-raid/parseAvatar.js';
 import { resolveDateLocale } from '../../../../functions/anti-raid/resolveDateLocale.js';
-import { insertAntiRaidNukeCaseLog } from '../../../../functions/logging/insertAntiRaidNukeCaseLog.js';
-import { upsertAntiRaidNukeReport } from '../../../../functions/logging/upsertGeneralLog.js';
 import type { AntiRaidNukeCommand } from '../../../../interactions/index.js';
-import { logger } from '../../../../logger.js';
-import { createButton } from '../../../../util/button.js';
-import { createMessageActionRow } from '../../../../util/messageActionRow.js';
 import { parseRegex } from '../../../../util/parseRegex.js';
 import { resolveTimestamp } from '../../../../util/timestamp.js';
 
@@ -108,8 +99,6 @@ export async function filter(
 			(!parsedConfusables.members || confusablesFilter(member)),
 	);
 
-	const pruneDays = Math.min(Math.max(Number(args.days ?? 1), 0), 7);
-
 	const parameterStrings = [];
 
 	if (parsedCreatedAfter || parsedCreatedBefore || parsedJoinAfter || parsedJoinBefore) {
@@ -176,150 +165,29 @@ export async function filter(
 		);
 	}
 
-	if (parameterStrings.length) {
-		parameterStrings.splice(
-			0,
-			0,
-			i18next.t('command.mod.anti_raid_nuke.common.parameters.heading', {
+	if (args.zalgo) {
+		parameterStrings.push(
+			i18next.t('command.mod.anti_raid_nuke.filter.parameters.zalgo', {
 				lng: locale,
 			}),
 		);
 	}
 
-	if (!members.size) {
-		await interaction.editReply({
-			content: `${i18next.t('command.mod.anti_raid_nuke.filter.errors.no_hits', {
-				lng: locale,
-			})}\n\n${parameterStrings.join('\n')}`,
-		});
-		return;
-	}
-
-	const banKey = nanoid();
-	const cancelKey = nanoid();
-	const dryRunKey = nanoid();
-
-	const banButton = createButton({
-		customId: banKey,
-		label: i18next.t('command.mod.anti_raid_nuke.common.buttons.execute', { lng: locale }),
-		style: ButtonStyle.Danger,
-	});
-	const cancelButton = createButton({
-		customId: cancelKey,
-		label: i18next.t('command.common.buttons.cancel', { lng: locale }),
-		style: ButtonStyle.Secondary,
-	});
-	const dryRunButton = createButton({
-		customId: dryRunKey,
-		label: i18next.t('command.mod.anti_raid_nuke.common.buttons.dry_run', { lng: locale }),
-		style: ButtonStyle.Primary,
-	});
-
-	const potentialHits = Buffer.from(members.map((member) => `${member.user.tag} (${member.id})`).join('\n'));
-	const potentialHitsDate = dayjs().format(DATE_FORMAT_LOGFILE);
-
-	const { creationRange, joinRange } = formatMemberTimestamps(members);
-
-	await interaction.editReply({
-		content: `${i18next.t('command.mod.anti_raid_nuke.common.pending', {
-			count: members.size,
-			creation_range: creationRange,
-			join_range: joinRange,
-			lng: locale,
-		})}\n\n${parameterStrings.join('\n')}`,
-		files: [{ name: `${potentialHitsDate}-anti-raid-nuke-list.txt`, attachment: potentialHits }],
-		components: [createMessageActionRow([cancelButton, banButton, dryRunButton])],
-	});
-
-	const collectedInteraction = await reply
-		.awaitMessageComponent({
-			filter: (collected) => collected.user.id === interaction.user.id,
-			componentType: ComponentType.Button,
-			time: 60000,
-		})
-		.catch(async () => {
-			try {
-				await interaction.editReply({
-					content: i18next.t('command.common.errors.timed_out', { lng: locale }),
-					components: [],
-				});
-			} catch (e) {
-				const error = e as Error;
-				logger.error(error, error.message);
-			}
-			return undefined;
-		});
-
-	if (collectedInteraction?.customId === cancelKey) {
-		await collectedInteraction.update({
-			content: i18next.t('command.mod.anti_raid_nuke.common.cancel', {
+	if (parsedConfusables.members) {
+		parameterStrings.push(
+			i18next.t('command.mod.anti_raid_nuke.filter.parameters.filter_confusables', {
 				lng: locale,
 			}),
-			components: [],
-		});
-	} else if (collectedInteraction?.customId === banKey || collectedInteraction?.customId === dryRunKey) {
-		const dryRunMode = collectedInteraction.customId === dryRunKey;
-
-		const content =
-			collectedInteraction.message.content +
-			(dryRunMode ? `\n\n${i18next.t('command.mod.anti_raid_nuke.common.parameters.dry_run', { lng: locale })}` : '');
-
-		await collectedInteraction.update({
-			content,
-			components: [
-				createMessageActionRow([
-					{ ...cancelButton, disabled: true },
-					{ ...banButton, disabled: true },
-				]),
-			],
-		});
-
-		const { result, cases } = await blastOff(
-			collectedInteraction,
-			{
-				days: pruneDays,
-				joinCutoff: parsedJoinAfter,
-				accountCutoff: parsedCreatedAfter,
-				dryRun: dryRunMode,
-			},
-			members,
-			locale,
 		);
+	}
 
-		if (!dryRunMode && cases.length) {
-			await insertAntiRaidNukeCaseLog(
-				collectedInteraction.guild,
-				collectedInteraction.user,
-				cases,
-				args.reason ??
-					i18next.t('command.mod.anti_raid_nuke.common.success', {
-						count: result.filter((r) => r.success).length,
-						lng: locale,
-					}),
-			);
-		}
-
-		const membersHit = Buffer.from(
-			result
-				.map(
-					(r) =>
-						`${r.member.user.id.padEnd(19, ' ')} | Join: ${dayjs(r.member.joinedTimestamp).format(
-							DATE_FORMAT_LOGFILE,
-						)} | Creation: ${dayjs(r.member.user.createdTimestamp).format(DATE_FORMAT_LOGFILE)} | ${r.member.user.tag}`,
-				)
-				.join('\n'),
-		);
-		const membersHitDate = dayjs().format(DATE_FORMAT_LOGFILE);
-
-		await upsertAntiRaidNukeReport(collectedInteraction.guild, collectedInteraction.user, result, dryRunMode);
-
-		await collectedInteraction.editReply({
-			content: i18next.t('command.mod.anti_raid_nuke.common.success', {
-				count: result.filter((r) => r.success).length,
+	if (parsedConfusables.pattern) {
+		parameterStrings.push(
+			i18next.t('command.mod.anti_raid_nuke.filter.parameters.clean_confusables', {
 				lng: locale,
 			}),
-			files: [{ name: `${membersHitDate}-anti-raid-nuke-hits.txt`, attachment: membersHit }],
-			components: [],
-		});
+		);
 	}
+
+	await handleAntiRaidNuke(interaction, reply, members, locale, parameterStrings, args.reason, args.days);
 }
