+++ packages/yuudachi/src/functions/logging/formatMembersToAttachment.ts
